using int64_t for comparators

+++ includes/sorts/ls_sorts.h
@@ -6,7 +6,7 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/22 12:54:49 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/11/30 12:49:16 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/12/08 16:39:47 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ extern void
 choose_sort(void *restrict base,
 	size_t n_el,
 	size_t width,
-	int32_t (*comparator)(const void*, const void*));
+	int64_t (*comparator)(const void*, const void*));
 
 /*
 ** quick sort
@@ -31,7 +31,7 @@ extern void
 q_sort(void *base,
 	const size_t n_el,
 	const size_t width,
-	int32_t (*comparator)(const void*, const void*));
+	int64_t (*comparator)(const void*, const void*));
 
 /*
 ** insertion sort
@@ -40,6 +40,6 @@ extern void
 i_sort(void *base,
 	const size_t n_el,
 	const size_t width,
-	int32_t (*comparator)(const void*, const void*));
+	int64_t (*comparator)(const void*, const void*));
 
 #endif
