+++ srcs/sorts/comparators/tabs/sort_ascii_tabs_cmp.c
@@ -6,13 +6,13 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/21 20:40:23 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/11/21 20:56:14 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/12/08 16:39:10 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ls.h"
 
-int32_t	sort_ascii_tabs_cmp(const void *a, const void *b)
+int64_t	sort_ascii_tabs_cmp(const void *a, const void *b)
 {
-	return (ft_strcmp(*(const char**)a, *(const char**)b));
+	return ((int64_t)ft_strcmp(*(const char**)a, *(const char**)b));
 }
